+++ packages/beacon-actorstore/src/index.test.ts
@@ -1,0 +1,6 @@
+describe("beacon-actorstore", () => {
+    it("needs tests", () => {
+        const calzone = "delicious";
+        expect(calzone).toBe("delicious");
+    });
+});