+++ packages/beacon-browserstorage/src/index.test.ts
@@ -1,0 +1,6 @@
+describe("beacon-browserstorage", () => {
+    it("needs tests", () => {
+        const stromboli = "delicious";
+        expect(stromboli).toBe("delicious");
+    });
+});