+++ packages/beacon-reactquery/src/withQuery.test.ts
@@ -1,31 +1,44 @@
-import { QueryClient } from "@tanstack/react-query";
-import { QueryStatus, withQuery } from "./withQuery";
-import { StoreConfig } from "@apogeelabs/beacon";
-
+/* eslint-disable @typescript-eslint/no-explicit-any */
 export type ProductQueryState = {
     products: any[];
+    isProductsLoading: boolean;
+    productsError: Error | null;
 };
 
 export type ProductQueryComputedState = {
     productCount: (state: ProductQueryState) => number;
 };
 
+// eslint-disable-next-line @typescript-eslint/no-empty-object-type
 export type ProductQueryActions = {};
 
 export type ProductQueryQueries = {
-    products: QueryStatus;
+    products: () => Promise<void>;
 };
+
+const mockRunInAction = jest.fn();
+jest.mock("mobx", () => {
+    return {
+        runInAction: mockRunInAction,
+    };
+});
+
+import { withQuery } from "./withQuery";
 
 describe("withQuery", () => {
     let mockApi: any,
         origOnStoreCreated: any,
         mockStore: any,
         mockQueryClient: any,
-        mockQueryCache: any;
+        mockQueryCache: any,
+        origConsoleErr: any;
 
     beforeEach(async () => {
         jest.clearAllMocks();
         jest.resetModules();
+        origConsoleErr = console.error;
+        console.error = jest.fn();
+        mockRunInAction.mockImplementation(cb => cb());
         mockStore = {};
         origOnStoreCreated = jest.fn();
         mockApi = {
@@ -39,6 +52,10 @@ describe("withQuery", () => {
             getQueryCache: jest.fn().mockReturnValue(mockQueryCache),
             fetchQuery: jest.fn().mockResolvedValue({}),
         };
+    });
+
+    afterEach(() => {
+        console.error = origConsoleErr;
     });
 
     describe("when the event doesn't have a query property", () => {
@@ -70,6 +87,8 @@ describe("withQuery", () => {
             const cfg = mwFn({
                 initialState: {
                     products: [],
+                    isProductsLoading: false,
+                    productsError: null,
                 },
                 derived: {
                     productCount: state => {
@@ -116,6 +135,8 @@ describe("withQuery", () => {
             const cfg = mwFn({
                 initialState: {
                     products: [],
+                    isProductsLoading: false,
+                    productsError: null,
                 },
                 derived: {
                     productCount: state => {
@@ -161,12 +182,18 @@ describe("withQuery", () => {
                         stateMapping: (_state, products) => {
                             return { products };
                         },
+                        statusMapping: {
+                            loading: "isProductsLoading",
+                            error: "productsError",
+                        },
                     },
                 },
             });
             const cfg = mwFn({
                 initialState: {
                     products: [],
+                    isProductsLoading: false,
+                    productsError: null,
                 },
                 derived: {
                     productCount: state => {
@@ -177,7 +204,14 @@ describe("withQuery", () => {
                 onStoreCreated: origOnStoreCreated,
             });
             cfg.onStoreCreated!(mockStore as any);
-            console.info("MOCK STORE", mockStore);
+        });
+
+        it("should call refetchQueries on the queryClient when the query is invokved", async () => {
+            await mockStore.queries.products();
+            expect(mockQueryClient.refetchQueries).toHaveBeenCalledTimes(1);
+            expect(mockQueryClient.refetchQueries).toHaveBeenCalledWith({
+                queryKey: ["products"],
+            });
         });
 
         it("should call the original onStoreCreated callback with the store instance", () => {
@@ -185,12 +219,9 @@ describe("withQuery", () => {
             expect(origOnStoreCreated).toHaveBeenCalledWith(mockStore);
         });
 
-        it("should call refetchQueries on the queryClient when refetch is invokved", async () => {
-            await mockStore.queries.products.refetch();
-            expect(mockQueryClient.refetchQueries).toHaveBeenCalledTimes(1);
-            expect(mockQueryClient.refetchQueries).toHaveBeenCalledWith({
-                queryKey: ["products"],
-            });
+        it("should set the state on the store", () => {
+            expect(mockStore.isProductsLoading).toEqual(true);
+            expect(mockStore.productsError).toEqual(undefined);
         });
     });
 
@@ -229,12 +260,22 @@ describe("withQuery", () => {
                             selector: (data: any) => {
                                 return data.items;
                             },
+                            statusMapping: {
+                                loading: (state: ProductQueryState, isLoading: boolean) => {
+                                    state.isProductsLoading = isLoading;
+                                },
+                                error: (state: ProductQueryState, err: Error | null) => {
+                                    state.productsError = err;
+                                },
+                            },
                         },
                     },
                 });
                 const cfg = mwFn({
                     initialState: {
                         products: [],
+                        isProductsLoading: false,
+                        productsError: null,
                     },
                     derived: {
                         productCount: state => {
@@ -245,7 +286,6 @@ describe("withQuery", () => {
                     onStoreCreated: origOnStoreCreated,
                 });
                 cfg.onStoreCreated!(mockStore as any);
-                console.info("MOCK STORE", mockStore);
             });
 
             it("should call the original onStoreCreated callback with the store instance", () => {
@@ -254,7 +294,7 @@ describe("withQuery", () => {
             });
 
             it("should call refetchQueries on the queryClient when refetch is invokved", async () => {
-                await mockStore.queries.products.refetch();
+                await mockStore.queries.products();
                 expect(mockQueryClient.refetchQueries).toHaveBeenCalledTimes(1);
                 expect(mockQueryClient.refetchQueries).toHaveBeenCalledWith({
                     queryKey: ["products"],
@@ -263,6 +303,261 @@ describe("withQuery", () => {
 
             it("should set the state on the store", () => {
                 expect(mockStore.products).toEqual(["ITEM1", "ITEM2"]);
+                expect(mockStore.isProductsLoading).toEqual(false);
+                expect(mockStore.productsError).toEqual(undefined);
+            });
+        });
+
+        describe("and it does not have a state selector", () => {
+            beforeEach(() => {
+                mockQueryCache.subscribe.mockImplementation((cb: any) => {
+                    cb({
+                        query: {
+                            options: {
+                                queryKey: ["products"],
+                            },
+                            state: {
+                                fetchStatus: "idle",
+                                data: {
+                                    items: ["ITEM1", "ITEM2"],
+                                },
+                            },
+                        },
+                    });
+                });
+                const mwFn = withQuery<
+                    ProductQueryQueries,
+                    ProductQueryState,
+                    ProductQueryComputedState,
+                    ProductQueryActions
+                >({
+                    queryClient: mockQueryClient,
+                    queries: {
+                        products: {
+                            queryKey: ["products"],
+                            queryFn: () => mockApi.fetchProducts(),
+                            stateMapping: (_state, products) => {
+                                return { products: products.items };
+                            },
+                            statusMapping: {
+                                loading: (state: ProductQueryState, isLoading: boolean) => {
+                                    state.isProductsLoading = isLoading;
+                                },
+                                error: (state: ProductQueryState, err: Error | null) => {
+                                    state.productsError = err;
+                                },
+                            },
+                        },
+                    },
+                });
+                const cfg = mwFn({
+                    initialState: {
+                        products: [],
+                        isProductsLoading: false,
+                        productsError: null,
+                    },
+                    derived: {
+                        productCount: state => {
+                            return state.products.length;
+                        },
+                    },
+                    actions: {},
+                    onStoreCreated: origOnStoreCreated,
+                });
+                cfg.onStoreCreated!(mockStore as any);
+            });
+
+            it("should call the original onStoreCreated callback with the store instance", () => {
+                expect(origOnStoreCreated).toHaveBeenCalledTimes(1);
+                expect(origOnStoreCreated).toHaveBeenCalledWith(mockStore);
+            });
+
+            it("should call refetchQueries on the queryClient when refetch is invokved", async () => {
+                await mockStore.queries.products();
+                expect(mockQueryClient.refetchQueries).toHaveBeenCalledTimes(1);
+                expect(mockQueryClient.refetchQueries).toHaveBeenCalledWith({
+                    queryKey: ["products"],
+                });
+            });
+
+            it("should set the state on the store", () => {
+                expect(mockStore.products).toEqual(["ITEM1", "ITEM2"]);
+                expect(mockStore.isProductsLoading).toEqual(false);
+                expect(mockStore.productsError).toEqual(undefined);
+            });
+        });
+    });
+
+    describe("when the query is erroring", () => {
+        beforeEach(() => {
+            mockQueryCache.subscribe.mockImplementation((cb: any) => {
+                cb({
+                    query: {
+                        options: {
+                            queryKey: ["products"],
+                        },
+                        state: {
+                            fetchStatus: "error",
+                            error: new Error("FETCH_ERROR"),
+                        },
+                    },
+                });
+            });
+            const mwFn = withQuery<
+                ProductQueryQueries,
+                ProductQueryState,
+                ProductQueryComputedState,
+                ProductQueryActions
+            >({
+                queryClient: mockQueryClient,
+                queries: {
+                    products: {
+                        queryKey: ["products"],
+                        queryFn: () => mockApi.fetchProducts(),
+                        stateMapping: (_state, products) => {
+                            return { products };
+                        },
+                        statusMapping: {
+                            loading: "isProductsLoading",
+                            error: "productsError",
+                        },
+                    },
+                },
+            });
+            const cfg = mwFn({
+                initialState: {
+                    products: [],
+                    isProductsLoading: false,
+                    productsError: null,
+                },
+                derived: {
+                    productCount: state => {
+                        return state.products.length;
+                    },
+                },
+                actions: {},
+                onStoreCreated: origOnStoreCreated,
+            });
+            cfg.onStoreCreated!(mockStore as any);
+        });
+
+        it("should call the original onStoreCreated callback with the store instance", () => {
+            expect(origOnStoreCreated).toHaveBeenCalledTimes(1);
+            expect(origOnStoreCreated).toHaveBeenCalledWith(mockStore);
+        });
+
+        it("should set the state on the store", () => {
+            expect(mockStore.isProductsLoading).toEqual(false);
+            expect(mockStore.productsError).toEqual(new Error("FETCH_ERROR"));
+        });
+    });
+
+    describe("when queryClient.fetchQuery rejects", () => {
+        describe("with string status mappings", () => {
+            beforeEach(() => {
+                mockQueryClient.fetchQuery.mockRejectedValue(new Error("FETCH_ERROR"));
+                const mwFn = withQuery<
+                    ProductQueryQueries,
+                    ProductQueryState,
+                    ProductQueryComputedState,
+                    ProductQueryActions
+                >({
+                    queryClient: mockQueryClient,
+                    queries: {
+                        products: {
+                            queryKey: ["products"],
+                            queryFn: () => mockApi.fetchProducts(),
+                            stateMapping: (_state, products) => {
+                                return { products };
+                            },
+                            statusMapping: {
+                                loading: "isProductsLoading",
+                                error: "productsError",
+                            },
+                        },
+                    },
+                });
+                const cfg = mwFn({
+                    initialState: {
+                        products: [],
+                        isProductsLoading: false,
+                        productsError: null,
+                    },
+                    derived: {
+                        productCount: state => {
+                            return state.products.length;
+                        },
+                    },
+                    actions: {},
+                    onStoreCreated: origOnStoreCreated,
+                });
+                cfg.onStoreCreated!(mockStore as any);
+            });
+
+            it("should set the state on the store", () => {
+                expect(mockStore.isProductsLoading).toEqual(false);
+                expect(mockStore.productsError).toEqual(new Error("FETCH_ERROR"));
+            });
+        });
+
+        describe("with callback status mappings (& with invoking the cleanup fn returned by the middleware)", () => {
+            let unsubscribe: any;
+
+            beforeEach(() => {
+                unsubscribe = jest.fn();
+                mockQueryCache.subscribe.mockReset();
+                mockQueryCache.subscribe.mockReturnValue(unsubscribe);
+                mockQueryClient.fetchQuery.mockRejectedValue("FETCH_ERROR");
+                const mwFn = withQuery<
+                    ProductQueryQueries,
+                    ProductQueryState,
+                    ProductQueryComputedState,
+                    ProductQueryActions
+                >({
+                    queryClient: mockQueryClient,
+                    queries: {
+                        products: {
+                            queryKey: ["products"],
+                            queryFn: () => mockApi.fetchProducts(),
+                            stateMapping: (_state, products) => {
+                                return { products };
+                            },
+                            statusMapping: {
+                                loading: (state: ProductQueryState, isLoading: boolean) => {
+                                    state.isProductsLoading = isLoading;
+                                },
+                                error: (state: ProductQueryState, err: Error | null) => {
+                                    state.productsError = err;
+                                },
+                            },
+                        },
+                    },
+                });
+                const cfg = mwFn({
+                    initialState: {
+                        products: [],
+                        isProductsLoading: false,
+                        productsError: null,
+                    },
+                    derived: {
+                        productCount: state => {
+                            return state.products.length;
+                        },
+                    },
+                    actions: {},
+                    onStoreCreated: origOnStoreCreated,
+                });
+                const mwCleanupFn = cfg.onStoreCreated!(mockStore as any);
+                (mwCleanupFn as any)();
+            });
+
+            it("should set the state on the store", () => {
+                expect(mockStore.isProductsLoading).toEqual(false);
+                expect(mockStore.productsError).toEqual(new Error("FETCH_ERROR"));
+            });
+
+            it("should unsubscribe from the query cache", () => {
+                expect(unsubscribe).toHaveBeenCalledTimes(1);
             });
         });
     });