+++ packages/beacon-browserstorage/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @apogeelabs/beacon-browserstorage
 
+## 0.3.0
+
+### Minor Changes
+
+- Continuing PoC work
+
+### Patch Changes
+
+- Updated dependencies
+    - @apogeelabs/beacon@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes